currency page moved to core and reused in both clients
