+++ clients/extension/src/navigation/router.tsx
@@ -5,7 +5,6 @@ import { ReshareSecureVault } from '@clients/extension/src/components/settings/r
 import { SetupFastVaultPage } from '@clients/extension/src/components/setup/SetupFastVaultPage'
 import { SetupSecureVaultPage } from '@clients/extension/src/components/setup/SetupSecureVaultPage'
 import { AppPath, appPaths } from '@clients/extension/src/navigation'
-import { CurrencyPage } from '@clients/extension/src/pages/popup/pages/currency'
 import DeleteVaultPage from '@clients/extension/src/pages/popup/pages/delete-vault'
 import ImportFilePage from '@clients/extension/src/pages/popup/pages/import-file'
 import ImportQRPage from '@clients/extension/src/pages/popup/pages/import-qr'
@@ -20,6 +19,7 @@ import { VaultSettingsPage } from '@clients/extension/src/pages/popup/pages/vaul
 import { VaultsPage } from '@clients/extension/src/pages/popup/pages/vaults'
 import { CorePath, corePaths } from '@core/ui/navigation'
 import { sharedRoutes } from '@core/ui/navigation/routes'
+import { CurrencyPage } from '@core/ui/preferences/currency/CurrencyPage'
 import { ActiveVaultGuard } from '@core/ui/vault/ActiveVaultGuard'
 import { toEntries } from '@lib/utils/record/toEntries'
 import { ReactNode } from 'react'
