+++ core/ui/preferences/currency/CurrencyPage.styles.ts
@@ -1,31 +1,24 @@
 import { UnstyledButton } from '@lib/ui/buttons/UnstyledButton'
 import { borderRadius } from '@lib/ui/css/borderRadius'
-import { PageSlice } from '@lib/ui/page/PageSlice'
 import { getColor } from '@lib/ui/theme/getters'
 import styled from 'styled-components'
 
 export const CurrencyBox = styled.div`
-  flex: 1;
-  display: flex;
-  flex-direction: column;
   align-items: flex-start;
-
-  font-weight: 400;
-  font-size: 16px;
   color: ${getColor('contrast')};
+  display: flex;
+  flex: 1;
+  flex-direction: column;
+  font-size: 16px;
+  font-weight: 400;
 `
 
 export const CurrencyButton = styled(UnstyledButton)`
+  align-items: center;
+  background-color: ${getColor('foreground')};
+  ${borderRadius.m};
   display: flex;
   justify-content: space-between;
-  align-items: center;
-
   padding: 16px;
-  ${borderRadius.m};
-  background-color: ${getColor('foreground')};
   width: 100%;
-`
-
-export const StyledPageSlice = styled(PageSlice)`
-  margin-bottom: 16px;
 `
